refactor(types): extract self-managed Connect types into dedicated file
Move ConnectAuthMethod, ConnectSaslScramAuth, ConnectTlsAuth, SelfManagedConnector,
SelfManagedConnectors, mergeSelfManagedConnectors, and SetSelfManagedConnectors out of
types.go and state.go into a new self_managed_connect.go inside the same internal/types
package. MSK Connect (ConnectorSummary, AWSClientInformation.Connectors) stays put. All
consumer imports use qualified types.X and continue to work without changes.

Enables tightening the kafka connect scan tests change_in trigger in a follow-up.
