+++ internal/types/types.go
@@ -73,25 +73,6 @@ func AllAuthTypes() []string {
 	}
 }
 
-type ConnectAuthMethod string
-
-const (
-	ConnectAuthMethodSaslScram       ConnectAuthMethod = "SASL/SCRAM"
-	ConnectAuthMethodTls             ConnectAuthMethod = "TLS"
-	ConnectAuthMethodUnauthenticated ConnectAuthMethod = "Unauthenticated"
-)
-
-type ConnectSaslScramAuth struct {
-	Username string
-	Password string
-}
-
-type ConnectTlsAuth struct {
-	CACert     string
-	ClientCert string
-	ClientKey  string
-}
-
 type MigrationType int
 
 const (
